requirements.txt
anyio==4.10.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
clip @ https://github.com/openai/CLIP/archive/d50d76daa670286dd6cacf3bcd80b5e4823fc8e1.zip#sha256=b5842c25da441d6c581b53a5c60e0c2127ebafe0f746f8e15561a006c6c3be6a
colorama==0.4.6
filelock==3.19.1
fsspec==2025.9.0
ftfy==6.3.1
google-auth==2.41.1
google-auth-oauthlib==1.2.2
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
oauthlib==3.3.1
open_clip_torch @ https://github.com/mlfoundations/open_clip/archive/bb6e834e9c70d9c27d0dc3ecedeebeaeb1ffad6b.zip#sha256=a594e737394d4463872151a1e84ab5fa502a4f72686943526b663fa939d786b1
packaging==25.0
pillow==11.3.0
protobuf==3.20.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.1
python-telegram-bot==22.5
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sentencepiece==0.2.1
sniffio==1.3.1
sympy==1.14.0
torch==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.13
